requirements.txt
astroid==2.4.2
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
idna==2.10
isort==5.4.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.6.0
python-dotenv==0.14.0
requests==2.24.0
rope==0.17.0
six==1.15.0
soupsieve==2.0.1
toml==0.10.1
urllib3==1.25.10
wrapt==1.12.1
